Retrofit isolation of loading time from reports
Following the logic for switching to ms timing reports, it's only wise
to do a similar clean up of the division of times. Some of the people
use independent programs for each part, others do not. The init times
can be added to each to show the run time of one part only to compare
with other, while the total time will compare to those who use one
program for the pair of parts.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>