Alter solutions for reporting full run time always
An earlier design choice was to skip all final reporting after part 1
was completed when there was no part 2 solution yet. With the split of
init time and part 1 processing time, I now need the combined total.

Changing the exit to a return when testing for part 2 reversed that
ancient decision.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>